post

+++ src/main/java/com/api/blog/controller/post/PostRestController.java
@@ -18,34 +18,40 @@ public class PostRestController {
     @Autowired
     private IPostService postService;
 
-    private static final Logger logger =
-            LoggerFactory.getLogger(PostRestController.class);
+    private static final Logger logger = LoggerFactory.getLogger(PostRestController.class);
 
     @GetMapping("/all")
-    public ResponseEntity<List<PostEntity>> findAllPost(){
+    public ResponseEntity<List<PostEntity>> findAllPost() {
         return ResponseEntity.ok().body(postService.findAllPosts());
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Optional<PostEntity>> findPostById(@PathVariable String id){
+    public ResponseEntity<Optional<PostEntity>> findPostById(@PathVariable String id) {
         return ResponseEntity.ok().body(postService.findPostById(id));
     }
 
     @PostMapping("/save")
-    public ResponseEntity<?> savePost(@RequestBody PostEntity post){
-        PostEntity newUps = new PostEntity ();
-        try{
-            newUps = postService.savePost(post);
-            logger.info("se creo el post ");
-        } catch (Exception e){
-            logger.error("error al registrar: " + e.getMessage());
-            return  ResponseEntity.internalServerError().build();
+    public ResponseEntity<?> savePost(@RequestHeader("X-User-Role") String role, @RequestBody PostEntity post) {
+        if (!"AUTOR".equals(role)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("No tienes permiso para agregar nuevos posts.");
         }
-        return new ResponseEntity<>(newUps, HttpStatus.CREATED);
+        PostEntity newPost;
+        try {
+            newPost = postService.savePost(post);
+            logger.info("Se creó el post");
+        } catch (Exception e) {
+            logger.error("Error al registrar: " + e.getMessage(), e);
+            return ResponseEntity.internalServerError().body("Error al registrar: " + e.getMessage());
+        }
+        return new ResponseEntity<>(newPost, HttpStatus.CREATED);
     }
 
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<?> deletePostById(@PathVariable String id){
+    public ResponseEntity<?> deletePostById(@RequestHeader("X-User-Role") String role, @PathVariable String id) {
+        if (!"ADMIN".equals(role)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("No tienes permiso para eliminar posts.");
+        }
+
         try {
             postService.deletePostById(id);
             return ResponseEntity.ok().build();
@@ -53,4 +59,5 @@ public class PostRestController {
             return ResponseEntity.status(500).body("Error al eliminar.");
         }
     }
+
 }
